% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_network.R
\name{handle_sir_data}
\alias{handle_sir_data}
\title{Handle SIR data from network models.}
\usage{
handle_sir_data(data, digits = 1)
}
\arguments{
\item{data}{A list of SIR model replicates, output of the \code{igraph::sir}
function.}

\item{digits}{Pass digits to round function.}
}
\value{
A data.table of SIR data.
}
\description{
Handle SIR data from network models.
}
