% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{model_case_2}
\alias{model_case_2}
\title{Runs an exploitation competition simulation returning structured output.}
\usage{
model_case_2(
  popsize,
  nItems,
  landsize,
  nClusters,
  clusterSpread,
  tmax,
  genmax,
  range_perception,
  handling_time,
  regen_time,
  nThreads,
  max_dispersal,
  mProb,
  mSize
)
}
\arguments{
\item{popsize}{The population size.}

\item{nItems}{How many food items on the landscape.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterSpread}{How dispersed food is around the cluster centre.}

\item{tmax}{The number of timesteps per generation.}

\item{genmax}{The maximum number of generations per simulation.}

\item{range_perception}{The movement range for agents.}

\item{handling_time}{The handling time.}

\item{regen_time}{The item regeneration time.}

\item{nThreads}{How many threads to parallelise over. Set to 1 to run on
the HPC Peregrine cluster.}

\item{max_dispersal}{The maximum dispersal distance.
Setting this to 10 percent is already almost equivalent to global dispersal.}

\item{mProb}{The probability of mutation. The suggested value is 0.01.
While high, this may be more appropriate for a small population; change this
value and \code{popsize} to test the simulation's sensitivity to these values.}

\item{mSize}{Controls the mutational step size, and represents the scale
parameter of a Cauchy distribution.}
}
\value{
An S4 class, \code{pathomove_output}, with simulation outcomes.
}
\description{
Run the simulation using parameters passed as
arguments to the corresponding R function.
}
