% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot networks from data from \code{simulation_output} objects.}
\usage{
plot_network(g)
}
\arguments{
\item{g}{A \code{tidygraph} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot networks from data from \code{simulation_output} objects.
}
